Refactor: 컴포넌트 분리 & 메타데이터  설정
