+++ app/auth/login/page.tsx
@@ -1,16 +1,22 @@
-import Button from '@/components/ui/Button';
 import Link from 'next/link';
 import SocialButton from '@/components/pages/auth/SocialButton';
+import BottomButton from '@/components/pages/auth/BottomButton';
+import { Metadata } from 'next';
 
-export default function Login() {
+export const metadata: Metadata = {
+  title: '로그인',
+  robots: 'noindex',
+};
+
+export default function LoginPage() {
   return (
     <>
-      <div className="flex flex-col items-center max-w-[330px]">
+      <div className="flex flex-col items-center max-w-[80%]">
         <h1 className="header1">로그인</h1>
-        <span className="text-center mt-4 mb-[50px] text-grey-600 body1">
+        <span className="text-center mt-4 mb-[50px] text-grey-600 body1 break-keep">
           SNS로 간편하게 로그인하고 더 많은 서비스를 즐겨보세요!
         </span>
-        <div className="flex gap-10 items-center mb-[82px]">
+        <div className="flex gap-6 tablet:gap-10 items-center mb-[38px] tablet:mb-[82px]">
           <SocialButton socialProvider="kakao" />
           <SocialButton socialProvider="naver" />
           <SocialButton socialProvider="google" hasBorder="border" />
@@ -24,15 +30,12 @@ export default function Login() {
           </Link>
         </div>
       </div>
-      <div className="flex flex-col items-center gap-10 w-full">
-        <Button
-          className="text-lg flex-1 px-[20px] py-[16px]"
-          variant="ghost"
-          fullWidth
-          to="/auth/register"
-        >
-          간편 회원가입 하기
-        </Button>
+      <BottomButton
+        text="간편 회원가입 하기"
+        to="/auth/register"
+        variant="ghost"
+        isFullWidth
+      >
         <div className="flex justify-between gap-[24px]">
           <Link href="/" className="text-grey-400 body1">
             아이디(이메일)찾기
@@ -42,7 +45,7 @@ export default function Login() {
             비밀번호 찾기
           </Link>
         </div>
-      </div>
+      </BottomButton>
     </>
   );
 }
