+++ app/auth/register/page.tsx
@@ -1,3 +1,10 @@
-export default function Register() {
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: '회원가입',
+  robots: 'noindex',
+};
+
+export default function RegisterPage() {
   return <div>Register</div>;
 }
