+++ components/ui/Input/SearchInput/index.tsx
@@ -17,7 +17,6 @@ export default function SearchInput({
   placeholder,
   className,
 }: SearchInputType) {
-  console.log(value);
   return (
     <div className="relative w-full">
       <input
